move ingress functions to ingress.go, fix some small issues based on review comments
